# contracts/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Sejong Civil AI API
  version: 3.0.0-draft
  description: |
    Final-scope draft. Validate through discovery/interview before implementation.
    Admin routes are local/private demo surfaces, not public authentication boundaries.
...
servers:
  - url: http://localhost:8000
paths:
...
  /api/v1/chat:
    post:
      operationId: createChatAnswer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChatRequest' }
      responses:
        '200':
          description: Success, follow-up, or safe fallback
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatResponse' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
...
  responses:
...
    ValidationError:
      description: Value-free request validation error without request input echo.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrorEnvelope' }
    ServiceUnavailable:
      description: No safe response can be produced from approved ACTIVE KB and currently available dependencies.
      headers:
        Retry-After:
          description: Suggested retry delay in seconds.
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ServiceUnavailableEnvelope' }
          example:
            error:
              code: SERVICE_UNAVAILABLE
              message: 잠시 후 다시 시도해 주세요.
              request_id: 7d444840-9dc0-11d1-b245-5ffdce74fad2
              retryable: true
  schemas:
...
    Intent:
      type: string
      enum: [MOVE_IN_RESIDENT_REGISTRATION, CERTIFICATE_ISSUANCE, BULKY_WASTE, LOCAL_TAX_GENERAL, OUT_OF_SCOPE, UNKNOWN]
...
    ChatAnswerStatus:
      type: string
      description: SYSTEM_ERROR is represented by the HTTP 503 envelope, not a 200 ChatResponse.
      enum: [SUCCESS, FOLLOWUP, FALLBACK]
    FallbackReason:
      type: string
      enum: [INSUFFICIENT_GROUNDING, PERSONAL_LOOKUP, LEGAL_JUDGMENT, OUT_OF_SCOPE, PRIVACY_UNRESOLVED]
...
    ServiceUnavailableEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id, retryable]
          properties:
            code: { const: SERVICE_UNAVAILABLE }
            message: { type: string, minLength: 1, maxLength: 200 }
            request_id: { type: string, format: uuid }
            retryable: { const: true }
    ValidationErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ValidationErrorDetail' }
    ValidationErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message, request_id, retryable]
      properties:
        code: { const: VALIDATION_ERROR }
        message: { const: 입력값을 확인해 주세요. }
        request_id: { type: string, format: uuid }
        retryable: { const: false }
...
    ChatRequest:
      type: object
      additionalProperties: false
      required: [question]
      properties:
        question: { type: string, minLength: 1, maxLength: 1000 }
        context_token:
          type: [string, 'null']
          minLength: 1
          maxLength: 2048
          description: >
            Optional signed conversational context. Clients treat it as opaque and
            keep it only in current-tab memory. Missing, null, expired, invalid,
            or unsupported tokens are treated as no context, not authentication
            failures. The token is never a source of official facts or authority.
        selected_region: { type: [string, 'null'], enum: [아름동, 도담동, 조치원읍, null] }
        simple_language: { type: boolean, default: false }
    Source:
      type: object
      additionalProperties: false
      required: [source_id, title, url, last_verified_at]
      properties:
        source_id: { type: string }
        title: { type: string }
        url: { type: string, format: uri, pattern: '^https://' }
        last_verified_at: { type: string, format: date }
        used_fields: { type: array, items: { type: string } }
    Office:
      type: object
      required: [id, region, office_name, address, phone, source_title, last_verified_at]
      properties:
        id: { type: string }
        region: { type: string }
        office_name: { type: string }
        address: { type: string }
        phone: { type: string }
        opening_hours: { type: [string, 'null'] }
        map_url: { type: [string, 'null'], format: uri, pattern: '^https://' }
        source_title: { type: string }
        source_url: { type: string, format: uri, pattern: '^https://' }
        last_verified_at: { type: string, format: date }
    FallbackPayloadBase:
      type: object
      required: [reason, title, message, candidate_eligible]
      properties:
        reason: { $ref: '#/components/schemas/FallbackReason' }
        title: { type: string }
        message: { type: string }
        next_actions: { type: array, items: { type: string } }
        candidate_eligible: { type: boolean }
        office: { anyOf: [ { $ref: '#/components/schemas/Office' }, { type: 'null' } ] }
    InsufficientGroundingFallback:
      type: object
      allOf:
        - { $ref: '#/components/schemas/FallbackPayloadBase' }
        - type: object
          required: [reason, candidate_eligible]
          properties:
            reason: { const: INSUFFICIENT_GROUNDING }
            candidate_eligible: { const: true }
      unevaluatedProperties: false
    PersonalLookupFallback:
      type: object
      allOf:
        - { $ref: '#/components/schemas/FallbackPayloadBase' }
        - type: object
          required: [reason, candidate_eligible]
          properties:
            reason: { const: PERSONAL_LOOKUP }
            candidate_eligible: { const: false }
      unevaluatedProperties: false
    LegalJudgmentFallback:
      type: object
      allOf:
        - { $ref: '#/components/schemas/FallbackPayloadBase' }
        - type: object
          required: [reason, candidate_eligible]
          properties:
            reason: { const: LEGAL_JUDGMENT }
            candidate_eligible: { const: false }
      unevaluatedProperties: false
    OutOfScopeFallback:
      type: object
      allOf:
        - { $ref: '#/components/schemas/FallbackPayloadBase' }
        - type: object
          required: [reason, candidate_eligible]
          properties:
            reason: { const: OUT_OF_SCOPE }
            candidate_eligible: { const: false }
      unevaluatedProperties: false
    PrivacyUnresolvedFallback:
      type: object
      allOf:
        - { $ref: '#/components/schemas/FallbackPayloadBase' }
        - type: object
          required: [reason, title, message, next_actions, candidate_eligible, office]
          properties:
            reason: { const: PRIVACY_UNRESOLVED }
            title: { const: 개인정보를 안전하게 처리하지 못했어요 }
            message: { const: 개인정보를 빼거나 표현을 바꿔서 다시 질문해 주세요. }
            next_actions:
              type: array
              items: { const: '이름, 주소, 전화번호, 접수번호 등을 적지 마세요.' }
              minItems: 1
              maxItems: 1
            candidate_eligible: { const: false }
            office: { type: 'null' }
      unevaluatedProperties: false
    Fallback:
      oneOf:
        - { $ref: '#/components/schemas/InsufficientGroundingFallback' }
        - { $ref: '#/components/schemas/PersonalLookupFallback' }
        - { $ref: '#/components/schemas/LegalJudgmentFallback' }
        - { $ref: '#/components/schemas/OutOfScopeFallback' }
        - { $ref: '#/components/schemas/PrivacyUnresolvedFallback' }
    ChatResponseBase:
      type: object
      required: [request_id, answer_status, intent, sources, context_token]
      properties:
        request_id: { type: string, format: uuid }
        answer_status: { $ref: '#/components/schemas/ChatAnswerStatus' }
        intent: { $ref: '#/components/schemas/Intent' }
        confidence: { type: [number, 'null'], minimum: 0, maximum: 1 }
        summary: { type: [string, 'null'] }
        procedure_steps: { type: array, items: { type: string } }
        required_documents: { type: array, items: { type: string } }
        processing_time: { type: [string, 'null'] }
        fee: { type: [string, 'null'] }
        department: { type: [string, 'null'] }
        sources: { type: array, items: { $ref: '#/components/schemas/Source' } }
        followup_options: { type: array, items: { type: string } }
        fallback: { anyOf: [ { $ref: '#/components/schemas/Fallback' }, { type: 'null' } ] }
        context_token:
          type: [string, 'null']
          minLength: 1
          maxLength: 2048
          description: >
            Fresh signed 15-minute context for current-tab memory. SUCCESS and
            FOLLOWUP may return a token; FALLBACK always returns null. Never log,
            persist, display, or use it as authentication.
    SuccessResponse:
      type: object
      allOf:
        - { $ref: '#/components/schemas/ChatResponseBase' }
        - type: object
          required: [answer_status, intent, sources, office]
          properties:
            answer_status: { const: SUCCESS }
            intent:
              enum: [MOVE_IN_RESIDENT_REGISTRATION, CERTIFICATE_ISSUANCE, BULKY_WASTE, LOCAL_TAX_GENERAL]
            sources: { type: array, minItems: 1, items: { $ref: '#/components/schemas/Source' } }
            office: { anyOf: [ { $ref: '#/components/schemas/Office' }, { type: 'null' } ] }
            fallback: { type: 'null' }
            followup_options: { type: array, maxItems: 0 }
      unevaluatedProperties: false
    FollowupResponse:
      type: object
      allOf:
        - { $ref: '#/components/schemas/ChatResponseBase' }
        - type: object
          required: [answer_status, intent, sources, followup_options, office]
          properties:
            answer_status: { const: FOLLOWUP }
            intent:
              enum: [MOVE_IN_RESIDENT_REGISTRATION, CERTIFICATE_ISSUANCE, BULKY_WASTE, LOCAL_TAX_GENERAL, UNKNOWN]
            summary: { type: 'null' }
            procedure_steps: { type: array, maxItems: 0 }
            required_documents: { type: array, maxItems: 0 }
            processing_time: { type: 'null' }
            fee: { type: 'null' }
            department: { type: 'null' }
            sources: { type: array, maxItems: 0 }
            office: { type: 'null' }
            followup_options: { type: array, minItems: 1, items: { type: string } }
            fallback: { type: 'null' }
      unevaluatedProperties: false
    FallbackResponseBase:
      allOf:
        - { $ref: '#/components/schemas/ChatResponseBase' }
        - type: object
          required: [answer_status, intent, sources, fallback, context_token]
          properties:
            answer_status: { const: FALLBACK }
            intent: { $ref: '#/components/schemas/Intent' }
            confidence: { type: 'null' }
            summary: { type: 'null' }
            procedure_steps: { type: array, maxItems: 0 }
            required_documents: { type: array, maxItems: 0 }
            processing_time: { type: 'null' }
            fee: { type: 'null' }
            department: { type: 'null' }
            sources: { type: array, maxItems: 0 }
            followup_options: { type: array, maxItems: 0 }
            fallback: { $ref: '#/components/schemas/Fallback' }
            context_token: { type: 'null' }
    InsufficientGroundingResponse:
      type: object
      allOf:
        - { $ref: '#/components/schemas/FallbackResponseBase' }
        - type: object
          required: [intent, fallback]
          properties:
            intent:
              enum: [MOVE_IN_RESIDENT_REGISTRATION, CERTIFICATE_ISSUANCE, BULKY_WASTE, LOCAL_TAX_GENERAL]
            fallback: { $ref: '#/components/schemas/InsufficientGroundingFallback' }
      unevaluatedProperties: false
    PersonalLookupResponse:
      type: object
      allOf:
        - { $ref: '#/components/schemas/FallbackResponseBase' }
        - type: object
          required: [intent, fallback]
          properties:
            intent:
              enum: [MOVE_IN_RESIDENT_REGISTRATION, CERTIFICATE_ISSUANCE, BULKY_WASTE, LOCAL_TAX_GENERAL]
            fallback: { $ref: '#/components/schemas/PersonalLookupFallback' }
      unevaluatedProperties: false
    LegalJudgmentResponse:
      type: object
      allOf:
        - { $ref: '#/components/schemas/FallbackResponseBase' }
        - type: object
          required: [intent, fallback]
          properties:
            intent:
              enum: [MOVE_IN_RESIDENT_REGISTRATION, CERTIFICATE_ISSUANCE, BULKY_WASTE, LOCAL_TAX_GENERAL]
            fallback: { $ref: '#/components/schemas/LegalJudgmentFallback' }
      unevaluatedProperties: false
    OutOfScopeResponse:
      type: object
      allOf:
        - { $ref: '#/components/schemas/FallbackResponseBase' }
        - type: object
          required: [intent, fallback]
          properties:
            intent: { const: OUT_OF_SCOPE }
            fallback: { $ref: '#/components/schemas/OutOfScopeFallback' }
      unevaluatedProperties: false
    PrivacyUnresolvedResponse:
      type: object
      allOf:
        - { $ref: '#/components/schemas/FallbackResponseBase' }
        - type: object
          required: [intent, confidence, fallback]
          properties:
            intent: { const: UNKNOWN }
            confidence: { type: 'null' }
            fallback: { $ref: '#/components/schemas/PrivacyUnresolvedFallback' }
      unevaluatedProperties: false
    FallbackResponse:
      oneOf:
        - { $ref: '#/components/schemas/InsufficientGroundingResponse' }
        - { $ref: '#/components/schemas/PersonalLookupResponse' }
        - { $ref: '#/components/schemas/LegalJudgmentResponse' }
        - { $ref: '#/components/schemas/OutOfScopeResponse' }
        - { $ref: '#/components/schemas/PrivacyUnresolvedResponse' }
    ChatResponse:
      oneOf:
        - { $ref: '#/components/schemas/SuccessResponse' }
        - { $ref: '#/components/schemas/FollowupResponse' }
        - { $ref: '#/components/schemas/FallbackResponse' }
...